Use the `needs` keyword to speed up the pipeline
As the stages Code quality, Dependency scanning and Test are just in place for
our own piece of mind the jobs can be run without the need of the stage to be
finished.

The `needs` keyword allows us to specify which jobs should be finished before
starting the jobs. In essence this is the same for all jobs: the building of
the cache, which is used by all other jobs.

See https://docs.gitlab.com/ee/ci/yaml/index.html#needs for more information.